inference fix

+++ ExpertSystem/FuzzyInference/FuzzifiedExamResult.cs
@@ -18,15 +18,15 @@ namespace FuzzyInference
         public FuzzifiedExamResult(ExamResult x)
         {
             int percentage = x.Result;
-            Average = 1 - Math.Abs(75 - percentage)/35;
+            Average = 1.0 - (double)Math.Abs(75 - percentage)/35.0;
             if (percentage > 50)
-                High = 1 - Math.Abs(100 - percentage) / 50;
+                High = 1.0 - (double)Math.Abs(100 - percentage) / 50.0;
             else
-                High = 0;
+                High = 0.0;
             if (percentage < 80)
-                Low = 1 - (percentage - 30) / 50;
+                Low = 1.0 - (double)(percentage - 30) / 50.0;
             else
-                Low = 0;
+                Low = 0.0;
             ExamName = x.ExamName;
             ExamLevel = x.ExamLevel;
         }
